moved restore_singletons to conftest.py
